booking-service/requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.1
psycopg2-binary==2.8.5
# Flask-Migrate==2.5.3
python-dotenv==0.12.0
Flask-SocketIO==4.3.0